- Make the Aware service a foreground service, instead of the accessibility services. This makes Aware service active at all times, keeping everyone happy'ish. There are still some corner cases because of Doze and Android O. More testing is needed.
